fix: address review — cursor -L, abort readers on timeout, shared useGitInstall
- Follow redirects in the cursor installer (curl -L), matching the
  other three entries; a CDN redirect would otherwise pipe an HTML
  response into bash.
- Run installer output readers in a JoinSet so the install timeout
  aborts them with the child — detached tasks kept emitting "running"
  events after the terminal "failed" payload while the killed child's
  pipes drained.
- Extract the duplicated git install/retry state (error derivation,
  installing flag, retry action) from DevToolsStatus and
  ProviderReadiness into a shared useGitInstall hook.
